реализовал полноценный презентер и заменил его кодом в мейне
